auth/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
pyjwt==2.8.0
sqlmodel==0.0.8
pg8000==1.30.3
python-multipart==0.0.6
email-validator==2.0.0


